Changes for V.4 and ANSI C
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1449 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/hack1.0.3.d/hack.fight.c
@@ -206,7 +206,7 @@ register thrown;
 
 	tmp += u.udaminc + dbon();
 	if(u.uswallow) {
-		if((tmp -= u.uswldtim) <= 0) {
+		if((int)(tmp -= u.uswldtim) <= 0) {
 			pline("Your arms are no longer able to hit.");
 			return(TRUE);
 		}
